chore: rename misleading useCase
