+++ react-client/src/domain/repository/pet/PetInfoRepository.ts
@@ -1,7 +1,7 @@
-import { PutPetInfoRequest } from "~/domain/useCases/pet/GetPetInfoUseCase";
+import { PutPetInfoRequest } from "~/domain/useCases/pet/PetInfoUseCase";
 import { PetModel } from "../../models/pet/PetModel";
 
-export interface GetPetInfoRepository {
+export interface PetInfoRepository {
   query(petId: string): Promise<PetModel | null>;
   mutate(data: PutPetInfoRequest): Promise<boolean>;
 }
