+++ react-client/src/domain/useCases/pet/PetInfoUseCase.test.ts
@@ -1,12 +1,12 @@
 import { MockHttpClient } from "~/domain/mocks/MockHttpClient";
 import { HttpClientRepository } from "~/domain/repository/HttpClientRepository";
-import { GetPetInfoUseCase } from "./GetPetInfoUseCase";
+import { PetInfoUseCase } from "./PetInfoUseCase";
 import getPetInfoMock from "./mocks/getPetInfoMock.json";
 
 // We don't care about the implementation while running those tests
 jest.mock("../PetPlaceHttpClientUseCase", () => {});
 
-describe("GetPetInfoUseCase", () => {
+describe("PetInfoUseCase", () => {
   describe("GET", () => {
     it("should return null when there is no data", async () => {
       const httpClient = new MockHttpClient({ data: null });
@@ -105,7 +105,7 @@ describe("GetPetInfoUseCase", () => {
 
 // Test helpers
 function makeSut(httpClient?: HttpClientRepository) {
-  return new GetPetInfoUseCase(
+  return new PetInfoUseCase(
     "token",
     httpClient ||
       new MockHttpClient({
