+++ react-client/src/domain/useCases/pet/PetInfoUseCase.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { HttpClientRepository } from "~/domain/repository/HttpClientRepository";
-import { GetPetInfoRepository } from "~/domain/repository/pet/GetPetInfoRepository";
+import { PetInfoRepository } from "~/domain/repository/pet/PetInfoRepository";
 import { PetModel } from "../../models/pet/PetModel";
 import { PetPlaceHttpClientUseCase } from "../PetPlaceHttpClientUseCase";
 import { parseData } from "../util/parseData";
@@ -18,7 +18,7 @@ const putServerSchema = z.object({
 
 export type PutPetInfoRequest = z.infer<typeof putServerSchema>;
 
-export class GetPetInfoUseCase implements GetPetInfoRepository {
+export class PetInfoUseCase implements PetInfoRepository {
   private httpClient: HttpClientRepository;
 
   constructor(authToken: string, httpClient?: HttpClientRepository) {
@@ -30,7 +30,7 @@ export class GetPetInfoUseCase implements GetPetInfoRepository {
   }
 
   private handleError = (error: unknown): null => {
-    console.error("GetPetInfoUseCase query error", error);
+    console.error("PetInfoUseCase query error", error);
     return null;
   };
 
@@ -63,7 +63,7 @@ export class GetPetInfoUseCase implements GetPetInfoRepository {
 
       return false;
     } catch (error) {
-      console.error("AccountDetailsUseCase mutation error", error);
+      console.error("PetInfoUseCase mutation error", error);
       return false;
     }
   };
